update methods in partice-auth

+++ particle-auth/lib/network/model/serialize_sol_transreqentity.dart
@@ -8,7 +8,7 @@ part 'serialize_sol_transreqentity.g.dart';
 class SerializeSOLTransReqEntity {
   late String sender;
   late String receiver;
-  late BigInt lamports;
+  late int lamports;
 
   SerializeSOLTransReqEntity();
